+++ tests/encoding/test_attrs_encoders.py
@@ -1,14 +1,15 @@
+from collections import deque
 from typing import Any
 
+import pytest
 from attrs import asdict, define, field, has
 
 from esmerald import Gateway, post
-from esmerald.encoders import Encoder, register_esmerald_encoder
+from esmerald.encoders import LILYA_ENCODER_TYPES, Encoder, register_esmerald_encoder
 from esmerald.testclient import create_client
 
 
 class AttrsEncoder(Encoder):
-
     def is_type(self, value: Any) -> bool:
         return has(value)
 
@@ -19,7 +20,14 @@ class AttrsEncoder(Encoder):
         return annotation(**value)
 
 
-register_esmerald_encoder(AttrsEncoder)
+@pytest.fixture(autouse=True, scope="function")
+def additional_encoders():
+    token = LILYA_ENCODER_TYPES.set(deque(LILYA_ENCODER_TYPES.get()))
+    try:
+        register_esmerald_encoder(AttrsEncoder)
+        yield
+    finally:
+        LILYA_ENCODER_TYPES.reset(token)
 
 
 @define
@@ -30,9 +38,9 @@ class AttrItem:
 
 
 def test_can_parse_attrs(test_app_client_factory):
-
     @post("/create")
     async def create(data: AttrItem) -> AttrItem:
+        assert type(LILYA_ENCODER_TYPES.get()[0]) is AttrsEncoder
         return data
 
     with create_client(routes=[Gateway(handler=create)]) as client:
@@ -44,7 +52,6 @@ def test_can_parse_attrs(test_app_client_factory):
 
 
 def test_can_parse_attrs_errors(test_app_client_factory):
-
     @define
     class Item:
         sku: str = field()